Added button graphics, changed background
I had to make the buttons square because if I made the buttons
transparent and round, java doesn't properly paint underneath it, and
there are massive paint glitches. I tried messing around by making the
wallpaper in a separate class, but it was really complicated. The
buttons also didn't show up, so I didn't update that. Finally, the
wallpaper is changed to what we had originally planned.