More Updates
Added icon for launcher (but for some reason its not aligned to the
other icons on the screen...).
Added code to CurrentPlayer (overloaded canMove and move), not sure how
to access the keyboard listener though since it's in gamewindow and
CurrentPlayer has no access.
Set focus for splashscreen (iffy, will check again later).
Other updates from in class.