Fixed Null Pointer Exception, Sound Icon Problems
See comment in MenuFrames for details.
Note: not sure how to toggle sound in SidePanel. Two scenarios
1. User already created an instance of audio by clicking settings before
playing (this case isn't as hard to deal with).
2. SidePanel needs to create a new instance (this might be do-able if we
check if it's created right as a level starts, and make an instance if
it hasn't been created yet).
Main issue: giving SidePanel access to SettingsPanel (tried creating a
return method to return the panel if it's been created, didn't work)