Icon Updates, Difficult Level, Ring, etc.
1. I updated the icons for easy game. Trying out a different style. Tell
me if you want to keep it.
2. The difficult level icons and game are created. Since all similar
charges are the same colour, when playing the game, it's easy to just
match the colours and not actually add things up. So, we could not match
the colours to confuse them.
3. The database has all the difficult level element names.
4. GameWindow now has access to difficult level.
5. Easy level has a win/lose screen. (should add buttons + highscore
option)
6. DifficultParticle gets the images from the name, rather than the
database. Since there are so many icons, it's just easier to do that.
7. Tried some different backgrounds. "universe.png" looks so cool. Maybe
we can discuss alternate background options.