Finally Won Difficult (Fixed Issue), New Icons
1. The installation program works on XP. I finally won difficult level
on my XP. There was an issue where the button still showed up to
continue so I removed it.
2. In the chm file, please update the instructions image. It still has
the old title.
3. I added additional possible desktop icons. I don't know why when I
install it, the shortcut is pixelated because it has a high resolution.
I prefer DesktopIcon3.ico, but you can choose which you like the best.
4. Are you printing the title page of the user manual? I'm printing the
back cover right? What should I add...